pdf_ingestion/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "pdf-ingestion"
version = "0.1.0"
description = "Production-ready PDF extraction service for the reconciliation product"
requires-python = ">=3.11"
dependencies = [
    "fastapi==0.115.6",
    "uvicorn[standard]==0.34.0",
    "pydantic==2.10.4",
    "pydantic-settings==2.7.1",
    "pdfplumber==0.11.4",
    "camelot-py[base]==0.11.0",
    "pikepdf==9.4.2",
    "structlog==24.4.0",
    "asyncpg==0.30.0",
    "alembic==1.14.1",
    "sqlalchemy[asyncio]==2.0.36",
    "boto3==1.35.86",
    "presidio-analyzer==2.2.355",
    "presidio-anonymizer==2.2.355",
    "httpx==0.28.1",
    "python-multipart==0.0.20",
    "PyJWT==2.9.0",
    "bcrypt==4.2.1",
]

[project.optional-dependencies]
dev = [
    "pytest==8.3.4",
    "pytest-asyncio==0.25.0",
    "hypothesis==6.119.3",
    "mypy==1.13.0",
    "ruff==0.8.6",
]

[tool.mypy]
strict = true
python_version = "3.11"
plugins = ["pydantic.mypy"]
exclude = ["tests/", "alembic/"]

[tool.pydantic-mypy]
init_forbid_extra = true
init_typed = true
warn_required_dynamic_aliases = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.ruff]
target-version = "py311"
line-length = 100
